+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2020-06-21
+
 ### Added
 
-- None
+- XML documentation on interfaces
+- Test project with unit tests and end-to-end service container registration tests
+- Dependency Injection component for service registration in the service container
+- `IDependencyInjectionRootComponent` extension `AddMapping` to attach component to root component
 
 ## [0.1.0] - 2020-06-20
 
@@ -19,7 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - README file describing project
 - Azure Pipelines based CI/CD setup
 - `IMapper` interface for implementing mappers
-- `IMappingService` to map using `IMapper`s in service provider
+- `IMappingService` and implementation `IMappingService` to map using `IMapper`s in service provider
 
-[unreleased]: https://github.com/SorenA/lightops-mapping/compare/0.1.0...develop
+[unreleased]: https://github.com/SorenA/lightops-mapping/compare/0.2.0...develop
+[0.2.0]: https://github.com/SorenA/lightops-mapping/tree/0.2.0
 [0.1.0]: https://github.com/SorenA/lightops-mapping/tree/0.1.0